Moving functions and correcting function names
